awe-agentic-web-explorer/requirements.txt
# AWE - Agentic Web Explorer
# Requirements

# Core dependencies
pydantic>=2.0.0
playwright>=1.40.0
httpx>=0.25.0

# Async support
aiofiles>=23.0.0

# LLM Providers
ollama>=0.1.0
groq>=0.4.0
openai>=1.10.0  # For Groq API compatibility

# Optional: Enhanced parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Better JSON handling
orjson>=3.9.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
